+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `AllocErrorType` trait shared between `alloc_mut` and `alloc` traits
+- `Error::Unsupported` variant for unsupported operations
 
 ### Changed